Bug fixes and enhancements to support the evaluation of feature/goal models with reused elements (not yet recursively evaluated), to properly set the contribution value of a reuse link, to auto-selected reused elements, to not allow more than one reuse link to be created per feature/goal reference, to ensure that a reuse link can only have a reused element as its source, to not allow a reused feature to be selected, to not allow a satisfaction value to be set for a reused goal model element, to not allow a contribution value of a reuse/mandatory/optional link to be set, to not allow toggling between contribution and correlation for reuse/mandatory/optional links, to support toggling between optional and mandatory links, to properly show the definition in the properties view of a reused element, to restrict changing the definition of a reused element through the properties view, to only allow features and not goal model elements to be selected as a new definition in the property view of a feature, and to fix a problem when deleting a reused feature in evaluation mode.
git-svn-id: svn://cserg0.site.uottawa.ca/projetseg/trunk@2903 1d5a7336-c65e-6947-a78c-d4236001c669

+++ src/seg/jUCMNav/model/commands/delete/DeleteGRLNodeCommand.java
@@ -12,6 +12,7 @@ import org.eclipse.gef.commands.CompoundCommand;
 import seg.jUCMNav.Messages;
 import seg.jUCMNav.model.commands.delete.internal.PreDeleteUrnModelElementCommand;
 import seg.jUCMNav.model.commands.delete.internal.RemoveURNmodelElementCommand;
+import seg.jUCMNav.strategies.util.ReusedElementUtil;
 import seg.jUCMNav.views.preferences.DeletePreferences;
 
 /**
@@ -44,8 +45,11 @@ public class DeleteGRLNodeCommand extends CompoundCommand {
 
             IntentionalElementRef intElement = (IntentionalElementRef) ref;
 
-            // Verify if this reference is the only one
-            if (intElement.getDef().getRefs().size() == 1 && DeletePreferences.getDeleteDefinition(ref)) {
+            // Verify if this reference is the only one and it is not a reused element since the
+            // definition should not be deleted (it is found in another model)
+            if (!ReusedElementUtil.isReusedElement(intElement.getDiagram().getUrndefinition().getUrnspec()
+            		.getGrlspec(), intElement.getDef()) && intElement.getDef().getRefs().size() == 1 
+            		&& DeletePreferences.getDeleteDefinition(ref)) {
                 // Add a command to delete the intentional element before execution
                 add(new DeleteIntentionalElementCommand(intElement.getDef()));
             }
